requirements.txt
fastapi
uvicorn[standard]
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
langgraph
langsmith
chromadb
pypdf
python-dotenv